update: convert Content.routes to tsx
